% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{item_bank_explorer}
\alias{item_bank_explorer}
\title{Launch item bank explorer}
\usage{
item_bank_explorer(item_bank)
}
\arguments{
\item{item_bank}{}
}
\value{

}
\description{
Launch item bank explorer
}
